server/requirements.txt
aiomysql==0.2.0
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==2.4.1
click==8.1.7
comm==0.2.2
databases==0.9.0
debugpy==1.8.8
decorator==5.1.1
executing==2.1.0
fastapi==0.115.4
greenlet==3.1.1
h11==0.14.0
idna==3.10
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.7.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mysql-connector-python==9.1.0
nest-asyncio==1.6.0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-multipart==0.0.17
pyzmq==26.2.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.41.2
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
uvicorn==0.32.0
wcwidth==0.2.13
mysql-connector-python 
python-dotenv
PyJWT==1.7.1